Fixed random kernel lock
When ipset is called to quickly, a kernel lock might prevent the
operation to execute successfully. The error is unfortunately system
dependend and might not appear for you at all, which makes it impossible
to test.

This *should* fix it for *most* systems hopefully. However, it slows
down starts and system tests.